+++ src/css/index.css
@@ -36,3 +36,8 @@ ul {
 li {
   list-style: none;
 }
+
+.lotto-header {
+  text-align: center;
+  margin: 3rem 0 2rem 0;
+}
