--- src/js/components/App.scss
@@ -1,4 +1,0 @@
-.lotto-header {
-  text-align: center;
-  margin: 3rem 0 2rem 0;
-}
