+++ src/js/components/purchasedLotto/PurchasedLotto.scss
