+++ src/js/components/purchasedLotto/lottoTicket/LottoTicket.js
