+++ src/js/components/purchasedLotto/lottoTicket/LottoTicket.scss
