+++ src/js/components/resultModal/ResultModal.scss
