+++ src/js/components/winningNumberForm/winningNumberInput/WinningNumberInput.scss
