+++ src/js/index.js
@@ -1,7 +1,7 @@
 import '../css/index.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
